move types to shared location
