+++ src/lib/tasks.ts
@@ -8,7 +8,6 @@ import process from "node:process";
 import { nodeInfo } from "./nodeInfo";
 import decompress from "decompress";
 import { Spinner } from "./spinner";
-import { bool } from "@polkadot/types";
 
 export async function checkCliDependencies(spinner: Spinner) {
   const dependencyList = [
@@ -57,7 +56,6 @@ export async function processTemplates(projectPath: string, templateData: Record
   });
 
   handlebars.registerHelper("if_eq", function (a, b, options): boolean {
-    console.log("AAAA", a, b, options);
     if (a === b) {
       // @ts-ignore
       return options.fn(this);
