+++ src/types/index.ts
@@ -10,3 +10,32 @@ export interface ChainProperty {
 }
 
 export type ExtrinsicPayload = SubmittableExtrinsic<"promise">;
+
+export type Encrypted = { iv: string; data: string };
+
+export interface AccountData {
+  isDev: boolean;
+  alias: string;
+  mnemonic: string | Encrypted;
+  address: string;
+}
+export interface SwankyConfig {
+  node: {
+    polkadotPalletVersions: string;
+    localPath: string;
+    supportedInk: string;
+  };
+  accounts: AccountData[];
+  contracts?: {
+    id: string;
+    language: "ask" | "ink";
+    name: string;
+    deployments: { artefactsPath: string; timestamp: number; address: string }[];
+    address: string;
+  }[];
+  networks: {
+    [network: string]: {
+      url: string;
+    };
+  };
+}
